+++ src/main/java/model/PlaceOrder.java
@@ -1,18 +1,18 @@
 package model;
 
 public class PlaceOrder{
-	private int petId;
+	private long petId;
 	private int quantity;
-	private int id;
+	private long id;
 	private String shipDate;
 	private boolean complete;
 	private String status;
 
-	public void setPetId(int petId){
+	public void setPetId(long petId){
 		this.petId = petId;
 	}
 
-	public int getPetId(){
+	public long getPetId(){
 		return petId;
 	}
 
@@ -24,11 +24,11 @@ public class PlaceOrder{
 		return quantity;
 	}
 
-	public void setId(int id){
+	public void setId(long id){
 		this.id = id;
 	}
 
-	public int getId(){
+	public long getId(){
 		return id;
 	}
